backend/requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
asyncmy==0.2.11
attrs==25.4.0
av==16.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
bleach==6.3.0
boto3==1.42.61
botocore==1.42.61
brotli==1.2.0
build==1.4.0
cachetools==7.0.2
cbor==1.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
clingo==5.8.0
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
doclayout_yolo==0.0.4
docling==2.76.0
docling-core==2.67.1
docling-ibm-models==3.11.0
docling-parse==5.5.0
docopt==0.6.2
docstring_parser==0.17.0
durationpy==0.10
einops==0.8.2
et_xmlfile==2.0.0
executing==2.2.1
Faker==40.8.0
fast-langdetect==0.2.5
fastapi==0.135.1
fastjsonschema==2.21.2
fasttext-predict==0.9.2.4
ffmpy==1.0.0
filelock==3.19.1
filetype==1.2.0
fire==0.7.1
FlagEmbedding==1.3.5
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
googleapis-common-protos==1.72.0
gradio==5.50.0
gradio_client==1.14.0
greenlet==3.3.2
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
hf_transfer==0.1.9
html5rdf==1.2.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-retries==0.4.6
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ijson==3.5.0
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
inscriptis==2.7.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
ir_datasets==0.5.11
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json_repair==0.58.4
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.10
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-deepseek==1.0.1
langchain-experimental==0.4.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.12
latex2mathml==3.78.1
lazy_loader==0.4
llamafactory==0.9.4
loguru==0.7.3
lxml==6.0.2
lz4==4.4.5
magika==1.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
milvus-model==0.2.12
mineru==2.7.6
mineru_vl_utils==0.1.22
mistune==3.2.0
mmh3==5.2.0
modelscope==1.34.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
mysql-connector-python==9.6.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
neo4j==5.28.3
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.24.2
openai==2.24.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
owlrl==7.1.4
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pdfminer.six==20260107
pdftext==0.6.3
peft==0.17.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.9.4
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
polyfactory==3.3.0
posthog==5.4.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.1
pybase64==1.4.3
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.3
pydantic-settings==2.13.1
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pylatexenc==2.10
pymilvus==2.6.9
pyparsing==3.3.2
pypdf==6.7.5
pypdfium2==4.30.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyshacl==0.31.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2026.1.post1
pyvis==0.3.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
rapidocr==3.7.0
rdflib==7.6.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
robust-downloader==0.0.2
rpds-py==0.30.0
rtree==1.4.1
ruff==0.15.4
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
semchunk==2.2.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
shtab==1.8.0
simsimd==6.5.15
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.2
stack-data==0.6.3
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchdata==0.11.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.56.1
trec-car-tools==2.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.6.0
trl==0.24.0
typer==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.3
ultralytics==8.4.19
ultralytics-thop==2.0.18
unlzw3==0.2.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarg==0.1.9
yarl==1.23.0
zipp==3.23.0
zlib-state==0.1.10
zstandard==0.25.0
umap-learn>=0.5.3